Make diagnostic sessions configurable
